# docs/openapi.json
{"openapi": "3.0.2", "info": {"title": "Parceiros API", "description": "Uma API para gerenciamento de parceiros, permitindo atualizar e criar novos parceiros a partir de arquivos CSV e listar os parceiros existentes.", "version": "1.0.0", "contact": {"name": "Paulo Henrique Raulino da Silva", "email": "phraulino@outlook.com"}, "license": {"name": "MIT", "url": "https://opensource.org/licenses/MIT"}}, "paths": {"/favicon.ico": {"get": {"tags": ["Padr\u00f5es"], "summary": "Favicon", "operationId": "favicon_favicon_ico_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}, "/": {"get": {"tags": ["Padr\u00f5es"], "summary": "Root", "operationId": "root__get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}, "/partners/": {"get": {"tags": ["Crud Parceiros", "Resposta Desafio"], "summary": "Listar parceiros", "description": "Recupera uma lista de parceiros, com pagina\u00e7\u00e3o.", "operationId": "listar_parceiros_partners__get", "parameters": [{"description": "N\u00famero de registros para pular na listagem", "required": false, "schema": {"title": "Skip", "type": "integer", "description": "N\u00famero de registros para pular na listagem", "default": 0}, "example": 5, "name": "skip", "in": "query"}, {"description": "Limite de registros a serem retornados na listagem", "required": false, "schema": {"title": "Limit", "type": "integer", "description": "Limite de registros a serem retornados na listagem", "default": 10}, "example": 10, "name": "limit", "in": "query"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"title": "Response Listar Parceiros Partners  Get", "type": "array", "items": {"$ref": "#/components/schemas/PartnerSchema"}}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "post": {"tags": ["Crud Parceiros"], "summary": "Criar parceiro", "description": "Cria um novo parceiro com as informa\u00e7\u00f5es fornecidas.", "operationId": "criar_parceiro_partners__post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/PartnerJsonSchema"}}}, "required": true}, "responses": {"201": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/PartnerSchema"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/partners/{cnpj}": {"put": {"tags": ["Crud Parceiros"], "summary": "Atualizar parceiro", "description": "Atualiza as informa\u00e7\u00f5es do parceiro com o CNPJ fornecido.", "operationId": "atualizar_parceiro_partners__cnpj__put", "parameters": [{"required": true, "schema": {"title": "Cnpj", "type": "string"}, "name": "cnpj", "in": "path"}], "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/PartnerUpdateSchema"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/PartnerSchema"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "delete": {"tags": ["Crud Parceiros"], "summary": "Deletar parceiro", "description": "Deleta o parceiro com o CNPJ fornecido.", "operationId": "deletar_parceiro_partners__cnpj__delete", "parameters": [{"required": true, "schema": {"title": "Cnpj", "type": "string"}, "name": "cnpj", "in": "path"}], "responses": {"204": {"description": "Successful Response"}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/search/{cnpj}": {"get": {"tags": ["Pesquisa Parceiros"], "summary": "Buscar parceiro por CNPJ", "description": "Recupera as informa\u00e7\u00f5es do parceiro com o CNPJ fornecido.", "operationId": "buscar_cnpj_parceiro_search__cnpj__get", "parameters": [{"required": true, "schema": {"title": "Cnpj", "type": "string"}, "example": "01234567891234", "name": "cnpj", "in": "path"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/PartnerSchema"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/search/": {"get": {"tags": ["Pesquisa Parceiros"], "summary": "Pesquisar parceiros", "description": "Realiza uma pesquisa nos parceiros com base nos crit\u00e9rios fornecidos, podendo buscar por qualquer informa\u00e7\u00e3o presente nas colunas.", "operationId": "pesquisar_parceiros_search__get", "parameters": [{"required": true, "schema": {"title": "Search Criteria", "type": "string"}, "example": "Empresa Exemplo", "name": "search_criteria", "in": "query"}, {"required": false, "schema": {"title": "Skip", "type": "integer", "default": 0}, "name": "skip", "in": "query"}, {"required": false, "schema": {"title": "Limit", "type": "integer", "default": 10}, "name": "limit", "in": "query"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"title": "Response Pesquisar Parceiros Search  Get", "type": "array", "items": {"$ref": "#/components/schemas/PartnerSchema"}}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/upload/partners/": {"post": {"tags": ["Upload Parceiros", "Resposta Desafio"], "summary": "Carregar parceiros via CSV", "description": "Carregar parceiros a partir de um arquivo CSV enviado. O arquivo deve ter o seguinte formato: CNPJ, Nome, E-mail e Telefone.Exemplo de conte\u00fado CSV:\n\nCNPJ, Raz\u00e3o Social, Nome Fantasia, Telefone, Email, CEP \n12.345.678/9123-45,Sol Eterno,Sol Eterno LTDA,(21) 98207-9901,teste@test.com,22783-115", "operationId": "carregar_parceiros_upload_partners__post", "requestBody": {"content": {"multipart/form-data": {"schema": {"$ref": "#/components/schemas/Body_carregar_parceiros_upload_partners__post"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}}, "components": {"schemas": {"Body_carregar_parceiros_upload_partners__post": {"title": "Body_carregar_parceiros_upload_partners__post", "required": ["file"], "type": "object", "properties": {"file": {"title": "File", "type": "string", "description": "Arquivo CSV contendo os dados dos parceiros a serem carregados.", "format": "binary"}}}, "HTTPValidationError": {"title": "HTTPValidationError", "type": "object", "properties": {"detail": {"title": "Detail", "type": "array", "items": {"$ref": "#/components/schemas/ValidationError"}}}}, "PartnerJsonSchema": {"title": "PartnerJsonSchema", "required": ["cep", "cnpj"], "type": "object", "properties": {"razao_social": {"title": "Razao Social", "type": "string", "example": "01234567891234"}, "nome_fantasia": {"title": "Nome Fantasia", "type": "string", "example": "Empresa Exemplo"}, "telefone": {"title": "Telefone", "type": "string", "example": "(12) 3456-7890"}, "email": {"title": "Email", "type": "string", "example": "exemplo@email.com"}, "cep": {"title": "Cep", "type": "string", "example": "12345-678"}, "cnpj": {"title": "Cnpj", "type": "string", "example": "01.234.567/8912-34"}}}, "PartnerSchema": {"title": "PartnerSchema", "required": ["cep", "cnpj", "id_parceiro", "zip_code_info", "last_update"], "type": "object", "properties": {"razao_social": {"title": "Razao Social", "type": "string", "example": "01234567891234"}, "nome_fantasia": {"title": "Nome Fantasia", "type": "string", "example": "Empresa Exemplo"}, "telefone": {"title": "Telefone", "type": "string", "example": "(12) 3456-7890"}, "email": {"title": "Email", "type": "string", "example": "exemplo@email.com"}, "cep": {"title": "Cep", "type": "string", "example": "12345-678"}, "cnpj": {"title": "Cnpj", "type": "string", "example": "01.234.567/8912-34"}, "id_parceiro": {"title": "Id Parceiro", "type": "string"}, "zip_code_info": {"$ref": "#/components/schemas/ZipCodeSchema"}, "last_update": {"title": "Last Update", "type": "string", "format": "date-time"}}}, "PartnerUpdateSchema": {"title": "PartnerUpdateSchema", "type": "object", "properties": {"razao_social": {"title": "Razao Social", "type": "string", "example": "01234567891234"}, "nome_fantasia": {"title": "Nome Fantasia", "type": "string", "example": "Empresa Exemplo"}, "telefone": {"title": "Telefone", "type": "string", "example": "(12) 3456-7890"}, "email": {"title": "Email", "type": "string", "example": "exemplo@email.com"}, "cep": {"title": "Cep", "type": "string"}}}, "ValidationError": {"title": "ValidationError", "required": ["loc", "msg", "type"], "type": "object", "properties": {"loc": {"title": "Location", "type": "array", "items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}}, "msg": {"title": "Message", "type": "string"}, "type": {"title": "Error Type", "type": "string"}}}, "ZipCodeSchema": {"title": "ZipCodeSchema", "required": ["cep", "bairro", "localidade", "uf", "last_update"], "type": "object", "properties": {"cep": {"title": "Cep", "type": "string"}, "bairro": {"title": "Bairro", "type": "string"}, "localidade": {"title": "Localidade", "type": "string"}, "uf": {"title": "Uf", "type": "string"}, "last_update": {"title": "Last Update", "type": "string", "format": "date-time"}}}}}}